Bifurcate interfaces module; Circumflex -> Interface
